+++ app/people/page.tsx
@@ -1,55 +1,13 @@
 "use client";
-import React, { useEffect, useState, useCallback, Fragment } from "react";
-import { useRouter, useSearchParams, usePathname } from "next/navigation";
+import React, { useEffect, useState, Fragment } from "react";
+import { useRouter, usePathname } from "next/navigation";
 import withApollo from "@/lib/withApollo";
-import { gql, useQuery, useLazyQuery } from "@apollo/client";
-import {UserType } from "../components/addAssignmentModal";
-import { DateTime, Interval } from "luxon";
+import {  useQuery } from "@apollo/client";
 import { Listbox, Transition } from "@headlessui/react";
 import { CheckIcon, ChevronDownIcon } from "@heroicons/react/20/solid";
-import { WorkWeek, WorkWeekType } from "../components/workWeek";
-import {
-	parseProjectDates,
-	parseWorkWeekDate,
-	AssignmentType,
-} from "../people/helperFunctions";
+import { UserType } from "./typeInterfaces";
+import { GET_USER_LIST } from "./gqlQueries";
 
-const GET_USER_ASSIGNMENTS = gql`
-	query getUserAssignments($selectedUserId: ID!) {
-		userAssignments(userId: $selectedUserId) {
-			id
-			startsOn
-			endsOn
-			status
-			assignedUser {
-				name
-				id
-			}
-			workWeeks {
-				id
-				actualHours
-				assignmentId
-				cweek
-				estimatedHours
-				year
-			}
-			project {
-				name
-				id
-				startsOn
-				endsOn
-			}
-		}
-	}
-`;
-const GET_USER_LIST = gql`
-	{
-		users {
-			id
-			name
-		}
-	}
-`;
 
 const PeopleView: React.FC = () => {
 	const [clientSide, setClientSide] = useState(false);
@@ -90,22 +48,21 @@ const PeopleView: React.FC = () => {
 						<div className="w-1/2 mr-2">
 							<Listbox.Label className="block text-sm font-medium leading-6 text-gray-900">
 								Person
-						
-							<div className="relative mt-2">
-								<Listbox.Button className="relative cursor-default rounded-md bg-white py-1.5 pl-3 pr-10 text-left text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 focus:outline-none focus:ring-2 focus:ring-indigo-500 sm:text-sm sm:leading-6">
-									<span className="flex items-center">
-										<span className="ml-3 block truncate">
-											{selectedUser.name}
+								<div className="relative mt-2">
+									<Listbox.Button className="relative cursor-default rounded-md bg-white py-1.5 pl-3 pr-10 text-left text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 focus:outline-none focus:ring-2 focus:ring-indigo-500 sm:text-sm sm:leading-6">
+										<span className="flex items-center">
+											<span className="ml-3 block truncate">
+												{selectedUser.name}
+											</span>
 										</span>
-									</span>
-									<span className="pointer-events-none absolute inset-y-0 right-0 ml-3 flex items-center pr-2">
-										<ChevronDownIcon
-											className="h-5 w-5 text-gray-400"
-											aria-hidden="true"
-										/>
-									</span>
-								</Listbox.Button>
-							</div>
+										<span className="pointer-events-none absolute inset-y-0 right-0 ml-3 flex items-center pr-2">
+											<ChevronDownIcon
+												className="h-5 w-5 text-gray-400"
+												aria-hidden="true"
+											/>
+										</span>
+									</Listbox.Button>
+								</div>
 							</Listbox.Label>
 							<Transition
 								show={open}
