+++ app/projects/page.tsx
@@ -1,43 +1,17 @@
 "use client";
 import React from "react";
 import withApollo from "@/lib/withApollo";
-import { gql, useQuery } from "@apollo/client";
+import { useQuery } from "@apollo/client";
 import { useEffect, useState } from "react";
-import { UserType } from "../components/addAssignmentModal";
-import { ProjectType } from "../components/addProjectModal";
-interface ClientType {
-	id: number;
-	name: string;
-	projects: [ProjectType];
-}
-
-const GET_DATA = gql`
-	{
-		clients {
-			id
-			name
-			projects {
-				name
-				status
-				paymentFrequency
-				users {
-					id
-					name
-					companies {
-						name
-					}
-				}
-			}
-		}
-	}
-`;
+import { AssignmentType, ClientType } from "../people/typeInterfaces";
+import { GET_PROJECT_DATA } from "../people/gqlQueries";
 
 const Projects: React.FC = () => {
 	const [clientSide, setClientSide] = useState(false);
 	useEffect(() => {
 		setClientSide(true);
 	}, []);
-	const { loading, error, data } = useQuery(GET_DATA, {
+	const { loading, error, data } = useQuery(GET_PROJECT_DATA, {
 		context: {
 			headers: {
 				cookie: clientSide ? document.cookie : null,
@@ -56,7 +30,7 @@ const Projects: React.FC = () => {
 						<div key={`${client.id} + ${client.name}`}>
 							<p className="font-black underline">{client.name}</p>
 							<ul>
-								{client.projects.map((project) => {
+								{client.projects?.map((project) => {
 									return (
 										<li key={`${project.name} + ${project.id}`}>
 											<span className="font-bold">
@@ -65,17 +39,33 @@ const Projects: React.FC = () => {
 											<ul className="p-3">
 												<li>Payment Frequency: {project.paymentFrequency}</li>
 												<li>Status: {project.status}</li>
+												{project.startsOn ? (
+													<li>Starts On: {project.startsOn} </li>
+												) : (
+													""
+												)}
+												{project.endsOn ? (
+													<li>Ends On: {project.endsOn} </li>
+												) : (
+													""
+												)}
 												<span className="font-bold underline">
 													Users Assigned To: {project.name}
 												</span>
 												<ul className="list-inside list-decimal border p-5">
-													{project.users.map((user: UserType) => {
-														return (
-															<li key={`${project.name} + ${user.name}`}>
-																{user.name}
-															</li>
-														);
-													})}
+													{project.assignments?.map(
+														(assignment: AssignmentType) => {
+															return (
+																<li
+																	key={`${project.name} + ${assignment.assignedUser.name}`}
+																>
+																	{assignment.assignedUser.name}
+																	<p>Starts: {assignment.startsOn}</p>
+																	<p>Ends: {assignment.endsOn}</p>
+																</li>
+															);
+														}
+													)}
 												</ul>
 											</ul>
 										</li>
